OBGM-47 Refactor OutboundStockMovementExcelImporter
